move logic out of functional and into function class
